+++ packages/sequence/java/ircam/jmax/editors/sequence/track/MidiTrackEditor.java
@@ -93,7 +93,8 @@ public class MidiTrackEditor extends TrackBaseEditor
 		int pitch;
 		for(Enumeration e = track.getTrackDataModel().getEvents(); e.hasMoreElements();)
 		{
-			pitch = ((Integer)((TrackEvent)e.nextElement()).getProperty("pitch")).intValue();
+			//pitch = ((Integer)((TrackEvent)e.nextElement()).getProperty("pitch")).intValue();
+			pitch = (int)((Double)((TrackEvent)e.nextElement()).getProperty("pitch")).doubleValue();
 			if(pitch > max) max = pitch;
 		}
 		if(viewMode == PIANOROLL_VIEW)
@@ -108,7 +109,8 @@ public class MidiTrackEditor extends TrackBaseEditor
 		int pitch;
 		for(Enumeration e = track.getTrackDataModel().getEvents(); e.hasMoreElements();)
 		{
-			pitch = ((Integer)((TrackEvent)e.nextElement()).getProperty("pitch")).intValue();
+			//pitch = ((Integer)((TrackEvent)e.nextElement()).getProperty("pitch")).intValue();
+			pitch = (int)((Double)((TrackEvent)e.nextElement()).getProperty("pitch")).doubleValue();
 			if(pitch < min) min = pitch;
 		}
 		if(viewMode == PIANOROLL_VIEW)
